Finish TAsk

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -7,7 +7,7 @@
  */
 int _strlen(char *s)
 {
-	int i = 0;
+	long int i = 0;
 
 	if (!s)
 	return (0);
